otp_generation_and_verify/requirements.txt
Flask
Flask-Cors
qiskit
qiskit-aer
PyJWT
python-dotenv
gunicorn
pysmtplib

